fix: Make new file and collect functions in three files
