+++ index.html
@@ -50,7 +50,7 @@
         <a href="/">
           <img src="images/landing/logo.svg" alt="로고" />
         </a>
-        <a href="/signin.html" class="cta login"> 로그인 </a>
+        <a href="./signin/index.html" class="cta login"> 로그인 </a>
       </div>
     </nav>
     <header>
@@ -60,7 +60,7 @@
           <span>쉽게 저장하고 </span>
           <span>관리해 보세요</span>
         </h1>
-        <a href="/signup.html" class="cta add"> 링크 추가하기 </a>
+        <a href="./signup/index.html" class="cta add"> 링크 추가하기 </a>
         <div class="hero-img"></div>
       </div>
     </header>
