+++ signin/signin_functions.js
@@ -1,0 +1,4 @@
+import { enterMessage } from "../functions/signin_verifyInput.js";
+import { changePasswordVision, isValidAccess } from "../functions/common.js";
+
+export { enterMessage, changePasswordVision, isValidAccess };
