--- signin_functions.js
@@ -1,5 +1,0 @@
-import { enterMessage } from "./functions/signin_verifyInput.js";
-import changePasswordVision from "./functions/password_visibility.js";
-import { isValidAccess } from "./functions/user_exist.js";
-
-export { enterMessage, changePasswordVision, isValidAccess };
