--- signup_functions.js
@@ -1,22 +1,0 @@
-import { enterEmailMessage, emailValid } from "./functions/signup_email.js";
-import {
-  enterPasswordMessage,
-  passwordValid,
-} from "./functions/signup_password.js";
-import {
-  enterPasswordCheckMessage,
-  checkValid,
-} from "./functions/signup_check.js";
-import changePasswordVision from "./functions/password_visibility.js";
-import { isValidAccess } from "./functions/user_exist.js";
-
-export {
-  enterEmailMessage,
-  enterPasswordMessage,
-  enterPasswordCheckMessage,
-  changePasswordVision,
-  emailValid,
-  passwordValid,
-  checkValid,
-  isValidAccess,
-};
